changed licensing to slightly more permissive MPL
The reason for the change in license for future versions is that
we want to make sure that the software provided will always
remain free, available, and that further improvements will be
contributed back to the open source community. However,
we understand that the virality of a GPL license is a concern for
professional developers. As such, we feel that MPL 2 strikes
a comfortable middle ground in this case.